Add files via upload
A Command line tool for working with JSON Web Tokens (JWT)

A utility designed for JSON Web Tokens (JWT) manipulation is available through the command line. This tool aids in secure authentication and session management. The server generates an encrypted "Access Token" containing the "user Id" and "expires in" information using the ACCESS_TOKEN_SECRET. Subsequently, it sends this "access Token" to the client-side browser. The browser then stores the "Access Token" locally and includes it in all future requests made to the server.